refactor (styled) : (component).styled.js 로 구분하기
