+++ src/components/Main/Main.styled.js
@@ -1,0 +1,16 @@
+import styled from "styled-components";
+
+export const S = {};
+
+S.Main = styled.main`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 3.2rem;
+  padding: 2rem 0;
+
+  @media screen and (min-width: 768px) {
+    gap: 2.4rem;
+    padding: 4rem 3.2rem;
+  }
+`
