+++ src/components/Nav/Logo.styled.js
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+export const S = {};
+
+S.ImgLogo = styled.img`
+  width: 10rem;
+  height: auto;
+
+  @media screen and (min-width: 768px) {
+    width: 13rem;
+  }
+`
