+++ src/pages/FolderPage.js
@@ -3,7 +3,7 @@ import HeaderSearch from "../components/Header/HeaderSearch"
 import Main from "../components/Main/Main"
 import Navigation from "../components/Nav/Navigation"
 import { useState } from "react";
-import S from "../components/styled";
+import { S } from "../components/styled";
 import LinkSection from "../components/Main/LinkSection";
 
 function FolderPage() {
